pyproject.toml
[tool.poetry]
name = "noptiera"
version = "0.1.0"
description = "Personal knowledge pipeline: ingest URLs, extract content, tag and search"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
trafilatura = "^2.0"
chromadb = "^0.6"
openai = "^1.0"
click = "^8.1"
python-frontmatter = "^1.1"
rich = "^13.0"
python-slugify = "^8.0"
python-telegram-bot = "^21.0"
python-dotenv = "^1.0"
fastmcp = "^2.1"

[tool.poetry.scripts]
noptiera = "cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
